+++ src/rules/html/index.js
@@ -3,21 +3,25 @@ import favicon from './favicon/rule.js'
 import imgAlt from './imgAlt/rule.js'
 import internalLinks from './internalLinks/rule.js'
 import lang from './lang/rule.js'
+import linkText from './linkText/rule.js'
 import maxOutboundLinks from './maxOutboundLinks/rule.js'
 import metaDescription from './metaDescription/rule.js'
 import metaViewport from './metaViewport/rule.js'
 import noVideo from './noVideo/rule.js'
+import tabIndex from './tabindex/rule.js'
 import title from './title/rule.js'
 
-export default [ 
+export default [
   brokenLinks,
   favicon,
   imgAlt,
   internalLinks,
   lang,
+  linkText,
   maxOutboundLinks,
   metaDescription,
   metaViewport,
   noVideo,
   title,
+  tabIndex,
 ]